requirements.txt
awesome-slugify==1.6.5
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
python-slugify==6.1.2